Move grid renderables to their own folder
